refactor: move JS to folders with meaning
