+++ app/scripts/config/config.js
