+++ app/scripts/config/map-style.js
