+++ app/scripts/libs/find-closest-city.js
