+++ app/scripts/libs/get-icon.js
