+++ app/scripts/libs/is-right-to-left.js
