+++ app/scripts/views/language-switch.js
