+++ app/scripts/views/sidebar.js
